Use long instead of int for saving player exp

+++ src/main/java/me/danjono/inventoryrollback/gui/BackupMenu.java
@@ -14,7 +14,7 @@ import java.util.UUID;
 
 public record BackupMenu(
         Player player, UUID uuid, LogType type, Instant timestamp, ItemStack[] inventory,
-        ItemStack[] enderChest, Location location, double health, int hunger, float saturation, int experience
+        ItemStack[] enderChest, Location location, double health, int hunger, float saturation, long experience
 ) {
 
     public Inventory showItems() {
